Added new function 'getOptimalBgSize' to StapleTracker class
As DFS is heavy operation, it is target for optimisations. In several sources it is stated that cv::dfs works faster if size is multiple of 2,3,5. To calculate optimal size, one can use 'cv::getOptimalDFTSize' function. Variable 'cf_response_size' stands for Correlation Filter size and calculated from 'bg_size', so we can calculate needed 'bg_size' for optimal DFS sizes. Algo is easy: brute-force. We are selecting range (or grid) and checking every pair of width and height to meet criteria. If there is no optimal sizes in range than initial value of 'bg_size' is sent.
Function added for optimisation purpose.