Optimize method TrackerTrain and separate logic in getSubwindow
As current solution of StapleTracker supports tracking without target scale,
so in TrackerTrain variables that were created in every call and copied to
vector were marked as static and inited only at the first frame. As sizes
of channels and bounding boxes are not changed, this improvement reduces
initialisation and copy operations.

Method getSubwindow in addition to croping an image performed image resizing.
So to save the meaning of the method the resize functionality was excluded from it.
Also as a result of excluding, the signature of the method has changed.

Also some minor changes:

- The variable firstFrame was removed from class, and so the signature
  of the TrackerTrain has changed. It was removed to keep class clean.

- A magic constant 28 was introduced as new variable FEATURE_CHANNELS.

- A separator between methods now is 2 spaces wide.