Add benchmarks to the trackerUpdate method
Added 2 types of benchmarks: 1st type based on OpenCV implementation and
2nd type based on hand-made things (raw loops).

Warning: result correctness method states that matrices are non-equal.
But in the console u can also see that data is equal. In debug mode if
u inspect the variable then u see that last bits are different, so
lets imagine that we have relaxed equality rules.