Adopt the trackerUpdate for multiusage
When static variables in the trackerUpdate are initiated they have fixed
width and length, but when tracker is inited with new target, it's sizes
can be different so something can go wrong. Because of that previous commit
was updated and one-time initialization became one-target initialization.
So when new target is captured static data in trackerUpdate at first frame
will be reinitialised.

The variable firstFrame is back for first frame initialisation.

Also method ensure_real changed to work with out-parameter instead of
returning new Mat. In debug mode this function also performs checks.